move widget button to menu bar
